address warning related to randomized_svd
Future versions of sklearn are going to change the default behavior of
the randomized_svd() function from deterministic with seed=0 to
non-deterministic.  To avoid surprising results, there is now a warning
that is issued if neither of these options is explicitly specified.

Presumably the new non-deterministic default is better in some way,
enough to justify the change.  But I don't know if it's better enough to
make a difference in this context.  So in the interest of backwards
compatibility, this commit just specifies the old, deterministic
behavior.

Signed-off-by: Kale Kundert <kale@thekunderts.net>