Fix meshgrid warning related to indexing argument
Torch is in the process of changing the default behavior of the
`meshgrid()` function from `indexing='ij'` to `indexing='xy'`, for
better compatibbility with numpy.  In the meantime, using `meshgrid()`
without specifying the indexing argument generates a warning.  This
commit explicitly specifies the previous default for all uses of
`meshgrid()` in the codebase.  This doesn't change any behavior, it just
silences the warning and makes the code robust against future changes to
the default value of this argument.

See pytorch/pytorch#50276 for more details.

Signed-off-by: Kale Kundert <kale@thekunderts.net>